requirements.txt
cvzone==1.5.6
opencv_python==4.7.0.68
Pillow==9.5.0
